requirements.txt
pyTelegramBotAPI~=3.8.1
img2pdf~=0.4.1
PyPDF2~=1.26.0
pytz~=2021.1
PyYAML~=5.4.1
mysql-connector-python~=8.0.25